refactor: rename tools/tasks to tools/doit for clarity
Rename the doit task definitions directory from tools/tasks/ to tools/doit/
to make it immediately obvious these are doit task definitions, consistent
with other tool-specific directories like tools/hooks/ai/.

- Rename tools/tasks/ → tools/doit/
- Update all imports in dodo.py, tests/test_templates.py, and github.py

Closes #170
