+++ dodo.py
@@ -1,20 +1,20 @@
 """Doit task runner configuration.
 
-This file imports all task definitions from the tools/tasks/ modules.
+This file imports all task definitions from the tools/doit/ modules.
 Each module contains related tasks grouped by functionality.
 """
 
 # Configuration
-from tools.tasks.base import DOIT_CONFIG  # noqa: F401
+from tools.doit.base import DOIT_CONFIG  # noqa: F401
 
 # Build and publish tasks
-from tools.tasks.build import (  # noqa: F401
+from tools.doit.build import (  # noqa: F401
     task_build,
     task_publish,
 )
 
 # Documentation tasks
-from tools.tasks.docs import (  # noqa: F401
+from tools.doit.docs import (  # noqa: F401
     task_docs_build,
     task_docs_deploy,
     task_docs_serve,
@@ -23,7 +23,7 @@ from tools.tasks.docs import (  # noqa: F401
 )
 
 # Git tasks
-from tools.tasks.git import (  # noqa: F401
+from tools.doit.git import (  # noqa: F401
     task_bump,
     task_changelog,
     task_commit,
@@ -32,20 +32,20 @@ from tools.tasks.git import (  # noqa: F401
 )
 
 # GitHub issue and PR tasks
-from tools.tasks.github import (  # noqa: F401
+from tools.doit.github import (  # noqa: F401
     task_issue,
     task_pr,
 )
 
 # Installation tasks
-from tools.tasks.install import (  # noqa: F401
+from tools.doit.install import (  # noqa: F401
     task_install,
     task_install_dev,
     task_install_direnv,
 )
 
 # Maintenance tasks
-from tools.tasks.maintenance import (  # noqa: F401
+from tools.doit.maintenance import (  # noqa: F401
     task_cleanup,
     task_completions,
     task_completions_install,
@@ -54,7 +54,7 @@ from tools.tasks.maintenance import (  # noqa: F401
 )
 
 # Code quality tasks
-from tools.tasks.quality import (  # noqa: F401
+from tools.doit.quality import (  # noqa: F401
     task_check,
     task_complexity,
     task_deadcode,
@@ -66,7 +66,7 @@ from tools.tasks.quality import (  # noqa: F401
 )
 
 # Release tasks
-from tools.tasks.release import (  # noqa: F401
+from tools.doit.release import (  # noqa: F401
     task_release,
     task_release_dev,
     task_release_pr,
@@ -74,14 +74,14 @@ from tools.tasks.release import (  # noqa: F401
 )
 
 # Security tasks
-from tools.tasks.security import (  # noqa: F401
+from tools.doit.security import (  # noqa: F401
     task_audit,
     task_licenses,
     task_security,
 )
 
 # Testing tasks
-from tools.tasks.testing import (  # noqa: F401
+from tools.doit.testing import (  # noqa: F401
     task_coverage,
     task_test,
 )
